@@ -1,3 +1,5 @@
+import 'dart:ffi';
+
 import 'package:flutter/material.dart';
 
 const _cardMargins = 15.0;
@@ -6,9 +8,9 @@ const _cardBorderRadius = 10.0;
 class ReusableCard extends StatelessWidget {
   ReusableCard({@required this.color, this.cardChild, this.onTap});
 
-  final Color color;
-  final Widget cardChild;
-  final Function onTap;
+  final Color? color;
+  final Widget? cardChild;
+  final VoidCallback? onTap;
 
   @override
   Widget build(BuildContext context) {
